refactors static navbar css
