--- src/assets/components/_navbar.scss
@@ -1,170 +1,0 @@
-.onboard {
-  li {
-    float: left;
-    padding-right: 30px;
-    font-family: "Lato", sans-serif;
-    font-weight: 400;
-    font-size: 20px;
-    cursor: pointer;
-    user-select: none;
-    color: $darkGray;
-    opacity: 0.7;
-  }
-
-  .navbar {
-    padding-top: 20px;
-    padding-bottom: 20px;
-    padding-left: 10%;
-    padding-right: 10%;
-    position: relative;
-  }
-
-  body {
-  }
-
-  .compName {
-    font-family: "Lato", sans-serif;
-    font-weight: 700;
-    font-size: 40px;
-    color: $companyName;
-  }
-
-  .focus {
-    color: $companyName;
-  }
-}
-
-.withSideNavbar {
-  display: flex;
-  height: 100vh;
-  // width: 100%;
-
-  a {
-    padding: 0px;
-    margin: 0px;
-  }
-
-  .navContainer {
-    &:hover {
-      .navTitle {
-        max-width: 350px;
-        transition: all 0.5s ease-in-out;
-        padding-left: 15px;
-        padding-right: 45px;
-        opacity: 1;
-      }
-    }
-
-    .navPadding {
-      width: 100%;
-      height: 100%;
-      background-color: $darkNavbar;
-    }
-
-    .roundAllCorners {
-      border-radius: 10px;
-    }
-
-    .roundLeftCorners {
-      border-radius: 10px 0px 0px 10px;
-    }
-
-    height: 100%;
-    padding: 0.5%;
-    padding-right: 0;
-
-    .navLinks {
-      display: flex;
-      flex-direction: column;
-      justify-content: space-around;
-    }
-
-    .active {
-      background-color: $turqois;
-      border-radius: 10px;
-      padding: 10px;
-      margin-left: 5px;
-      margin-right: 5px;
-    }
-
-    .inactive {
-      padding-left: 15px;
-      padding-right: 15px;
-
-      border-radius: 10px;
-      padding: 10px;
-      margin-left: 5px;
-      margin-right: 5px;
-    }
-
-    .navItem {
-      color: white;
-      margin-top: 5px;
-      margin-bottom: 5px;
-      max-height: 44px;
-      cursor: pointer;
-      user-select: none;
-    }
-
-    .navTitle {
-      max-width: 0px;
-      opacity: 0;
-      transition: all 0.5s ease-in-out;
-      white-space: nowrap;
-      padding-left: 0px;
-      padding-right: 0px;
-      vertical-align: middle;
-      font-family: "Lato", sans-serif;
-      font-weight: 300;
-      letter-spacing: 1px;
-    }
-
-    .noCollapse {
-      .navTitle {
-        max-width: 240px;
-        padding-left: 15px;
-        padding-right: 15px;
-      }
-    }
-
-    .navIcon {
-      width: 24px;
-      height: 24px;
-      float: left;
-      // background-size: 24px;
-    }
-    .dashboard {
-      background-image: url("../icons/home.svg");
-    }
-
-    .sites {
-      background-image: url("../icons/site.svg");
-    }
-
-    .templates {
-      background-image: url("../icons/hvac.svg");
-    }
-
-    .users {
-      background-image: url("../icons/accountManagement.svg");
-    }
-
-    .settings {
-      background-image: url("../icons/settings.svg");
-    }
-
-    .logout {
-      background-image: url("../icons/logout.svg");
-    }
-
-    .logoutItem {
-      position: absolute;
-      bottom: 2%;
-    }
-  }
-
-  .contentContainer {
-    padding-left: 10%;
-    padding-right: 5%;
-  }
-}
