+++ src/assets/components/_onboard.scss
@@ -74,4 +74,37 @@
     text-align: center;
     margin-top: 15px;
   }
+  li {
+    float: left;
+    padding-right: 30px;
+    font-family: "Lato", sans-serif;
+    font-weight: 400;
+    font-size: 20px;
+    cursor: pointer;
+    user-select: none;
+    color: $darkGray;
+    opacity: 0.7;
+  }
+
+  .navbar {
+    padding-top: 20px;
+    padding-bottom: 20px;
+    padding-left: 10%;
+    padding-right: 10%;
+    position: relative;
+  }
+
+  body {
+  }
+
+  .compName {
+    font-family: "Lato", sans-serif;
+    font-weight: 700;
+    font-size: 40px;
+    color: $companyName;
+  }
+
+  .focus {
+    color: $companyName;
+  }
 }
