+++ src/assets/components/_topRoutes.scss
@@ -1,0 +1,4 @@
+.topRoutes {
+  display: flex;
+  height: 100vh;
+}
