+++ src/assets/styles.scss
@@ -7,12 +7,12 @@
 
 @import "/components/_contactUs";
 @import "/components/_requestAccount";
-@import "/components/_navbar";
+@import "/components/_staticNavbar";
 @import "/components/_dynamicNavbar";
 @import "/components/_onboard";
 @import "/components/_authCheck";
 @import "/components/_content";
-
+@import "/components/_topRoutes";
 @import "/components/_dashboard";
 @import "/components/_sites";
 @import "/components/_channelTemplates";
